backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pypdf==4.3.1
langchain-text-splitters==0.2.2
huggingface_hub>=0.26.0
numpy==1.26.4
